Added  an adjustable temperature that returns a velocity based on a boltzmann distribution - Using std::gamma_distribution with alpha=3/2 and beta is the boltzmann factor to mimick a boltzman velocity distribution - Added a get_random function that returns a random float from a min,max interval - Did a bit of cleanup on lingering warnings

+++ CPP_2D_3D/spheres.h
@@ -51,7 +51,7 @@ class Sphere
 };
 
 //std::vector<Sphere> initialize_spheres(int numSpheres);
-void initialize_spheres(std::vector<Sphere*>&, int numSpheres, bool is3D);
+void initialize_spheres(std::vector<Sphere*>&, long unsigned int numSpheres, bool is3D);
